requirements.txt
black
cryptography
flake8
flwr==1.4.0
isort
mypy
nltk
numpy
opacus
pandas
portalocker
pre-commit
pycryptodome
pyproject-flake8==5.0.4
pytest
pytest-cov
scikit-learn
tensorflow-privacy
toml
torch
torchdata
torcheval
torchinfo
torchtext
torchvision
types-protobuf
types-PyYAML
types-requests
types-setuptools
types-six
types-tabulate
wandb
